Fixed the typedef of unsigned long long.

+++ clock.cpp
@@ -14,43 +14,43 @@
 
 Clock::Clock(){
     NanoSecondSharedMemoryFD = shm_open(NanoSeconds, O_CREAT | O_RDWR, 0666);
-    ftruncate(NanoSecondSharedMemoryFD, sizeof(unsigned long long));
-    NanoSecondSharedMemoryPointer = mmap(0, sizeof(unsigned long long), PROT_READ | PROT_WRITE, MAP_SHARED, NanoSecondSharedMemoryFD, 0);
+    ftruncate(NanoSecondSharedMemoryFD, sizeof(ull));
+    NanoSecondSharedMemoryPointer = mmap(0, sizeof(ull), PROT_READ | PROT_WRITE, MAP_SHARED, NanoSecondSharedMemoryFD, 0);
     time = 0;
 }
 Clock::~Clock(){
     
 }
 void Clock::init(){
-    memcpy(NanoSecondSharedMemoryPointer, &time, sizeof(unsigned long long));
+    memcpy(NanoSecondSharedMemoryPointer, &time, sizeof(ull));
 }
 
 void Clock::incrementClock(){
     time += increment;
-    memcpy(NanoSecondSharedMemoryPointer, &time, sizeof(unsigned long long));
+    memcpy(NanoSecondSharedMemoryPointer, &time, sizeof(ull));
 }
-unsigned long long Clock::getTime(){
+ull Clock::getTime(){
     return time;
 }
 
 int Clock::getSeconds(){
     return time / BILLION;
 }
-unsigned long long Clock::getNanoSeconds(){
+ull Clock::getNanoSeconds(){
     return time % BILLION;
 }
 void Clock::unlink(){
-    munmap(NanoSecondSharedMemoryPointer, sizeof(unsigned long long));
+    munmap(NanoSecondSharedMemoryPointer, sizeof(ull));
     close(NanoSecondSharedMemoryFD);
     shm_unlink(NanoSeconds);
 }
 void Clock::unmap(){
     close(NanoSecondSharedMemoryFD);
-    munmap(NanoSecondSharedMemoryPointer, sizeof(unsigned long long));
+    munmap(NanoSecondSharedMemoryPointer, sizeof(ull));
 }
 
 void Clock::update(){
-    memcpy(&time, NanoSecondSharedMemoryPointer, sizeof(unsigned long long));
+    memcpy(&time, NanoSecondSharedMemoryPointer, sizeof(ull));
 }
 void Clock::setIncrement(int CurrentChildren){
     if(CurrentChildren == 0)
